refactor: model que representa computador
